[FE] - feat: finish api get leaderboards for contest

+++ core-service/core-domain/core-application-service/src/main/java/com/backend/programming/learning/system/core/service/domain/dto/method/query/contest/QueryLeaderboardOfContestResponse.java
@@ -14,6 +14,6 @@ public class QueryLeaderboardOfContestResponse {
     private final ContestUserResponseEntity participantRank;
     private final List<ContestUserResponseEntity> contestLeaderboard;
     private final int currentPage;
-    private final int totalItems;
+    private final long totalItems;
     private final int totalPages;
 }
